<fix> : warning C4244: 'argument' : conversion from 'std::streamsize' to 'unsigned int', possible loss of data

+++ WinSFXMakerDlg.cpp
@@ -236,7 +236,7 @@ BOOL CWinSFXMakerDlg::CreateSFXFile(LPCTSTR templatePath, LPCTSTR zipPath, LPCTS
 	}
 	std::streamsize zipSize = (UINT)zipFile.tellg();
 	zipFile.seekg(0, std::ios::beg);
-	std::vector<char> zipData(zipSize);
+	std::vector<char> zipData((int)zipSize);
 	zipFile.read(zipData.data(), zipSize);
 	zipFile.close();
